backend/requirements.txt
# FastAPI and Web
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
httpx==0.26.0
aiohttp==3.9.3

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Cache
redis==5.0.1
aioredis==2.0.1

# AI/ML
openai==1.12.0
anthropic==0.18.1
tiktoken==0.6.0

# Vector Databases
pinecone-client==3.0.2
weaviate-client==4.4.4

# Document Parsing
pypdf==4.0.1
python-docx==1.1.0

# Queue
celery[redis]==5.3.6

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Configuration
pydantic-settings==2.1.0

# Logging
structlog==24.1.0

# Retry Logic
tenacity==8.2.3

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0

# Code Quality
black==24.1.1
ruff==0.2.1
mypy==1.8.0
